work on Lab 7
