--- src/labs/lab8s.jl
@@ -1,155 +1,0 @@
-# This last problem is advanced so may be considered optional. 
-
-# An alternative to using reflections to introduce zeros is to use rotations.
-# This is particularly convenient for tridiagonal matrices, where one needs to only
-# make one sub-diagonal zero. Here we explore a tridiagonal QR built from rotations
-# in a way that the factorisation can be computed in $O(n)$ operations.
-
-
-# **Problem 4.1** Complete the implementation of `Rotations`, which represents an orthogonal matrix `Q` that is a product
-# of rotations of angle `θ[k]`, each acting on the entries `k:k+1`. That is, it returns $Q = Q_1⋯Q_k$ such that
-# $$
-# Q_k[k:k+1,k:k+1] = 
-# \begin{bmatrix}
-# \cos θ[k] & -\sin θ[k]\\
-# \sin θ[k] & \cos θ[k]
-# \end{bmatrix}
-# $$
-
-struct Rotations{T} <: AbstractMatrix{T}
-    θ::Vector{T} # a vector of angles
-end
-
-import Base: *, size, getindex
-
-## we use the number of rotations to deduce the dimensions of the matrix
-size(Q::Rotations) = (length(Q.θ)+1, length(Q.θ)+1)
-
-function *(Q::Rotations, x::AbstractVector)
-    T = promote_type(eltype(Q), eltype(x))
-    y = Vector{T}(x) # copies x to a new Vector whose eltype is T
-    ## TODO: Apply Q in O(n) operations, modifying y in-place
-
-    ## SOLUTION
-    θ = Q.θ
-    ## Does Q1....Qn x
-    for k = length(θ):-1:1
-        #below has 4 ops to make the matrix and 12 to do the matrix-vector multiplication,
-        #total operations will be 48n = O(n)
-        c, s = cos(θ[k]), sin(θ[k])
-        y[k:(k+1)] = [c -s; s c] * y[k:(k+1)]
-    end
-    ## END
-
-    y
-end
-
-function getindex(Q::Rotations, k::Int, j::Int)
-    ## TODO: Return Q[k,j] in O(n) operations (hint: use *)
-
-    ## SOLUTION
-    ## recall that A_kj = e_k'*A*e_j for any matrix A
-    ## so if we use * above, this will take O(n) operations
-    n = size(Q)[1]
-    ej = zeros(eltype(Q), n)
-    ej[j] = 1
-    ## note, must be careful to ensure that ej is a VECTOR
-    ## not a MATRIX, otherwise * above will not be used
-    Qj = Q * ej
-    Qj[k]
-    ## END
-end
-
-θ1 = randn(5)
-Q = Rotations(θ1)
-@test Q'Q ≈ I
-@test Rotations([π/2, -π/2]) ≈ [0 0 -1; 1 0 0; 0 -1 0]
-
-# When one computes a tridiagonal QR, we introduce entries in the
-# second super-diagonal. Thus we will use the `UpperTridiagonal` type
-# from Lab 4:
-
-import Base: *, size, getindex, setindex!
-struct UpperTridiagonal{T} <: AbstractMatrix{T}
-    d::Vector{T}   # diagonal entries
-    du::Vector{T}  # super-diagonal enries
-    du2::Vector{T} # second-super-diagonal entries
-end
-
-size(U::UpperTridiagonal) = (length(U.d),length(U.d))
-
-function getindex(U::UpperTridiagonal, k::Int, j::Int)
-    d,du,du2 = U.d,U.du,U.du2
-    if j - k == 0
-        d[j]
-    elseif j - k == 1
-        du[k]
-    elseif j - k == 2
-        du2[k]
-    else
-        0
-    end
-end
-
-function setindex!(U::UpperTridiagonal, v, k::Int, j::Int)
-    d,du,du2 = U.d,U.du,U.du2
-    if j > k+2
-        error("Cannot modify off-band")
-    end
-    if j - k == 0
-        d[k] = v
-    elseif j - k == 1
-        du[k] = v
-    elseif j - k == 2
-        du2[k] = v
-    else
-        error("Cannot modify off-band")
-    end
-    U # by convention we return the matrix
-end
-
-
-# **Problem 4.2** Combine `Rotations` and `UpperTridiagonal` from last problem sheet
-# to implement a banded QR decomposition, `bandedqr`, that only takes $O(n)$ operations. Hint: the
-# `atan(y,x)` function gives the angle of a vector `[x,y]`.
-
-
-function bandedqr(A::Tridiagonal)
-    n = size(A, 1)
-    Q = Rotations(zeros(n - 1)) # Assume Float64
-    R = UpperTridiagonal(zeros(n), zeros(n - 1), zeros(n - 2))
-
-    ## TODO: Populate Q and R by looping through the columns of A.
-
-    ## SOLUTION
-    R[1, 1:2] = A[1, 1:2]
-        
-    for j = 1:n-1
-        ## angle of rotation
-        Q.θ[j] = atan(A[j+1, j], R[j, j])
-        θ = -Q.θ[j] # rotate in opposite direction 
-
-        c, s = cos(θ), sin(θ)
-        ## [c -s; s c] represents the rotation that introduces a zero.
-        ## This is [c -s; s c] to j-th column, but ignore second row
-        ## which is zero
-        R[j, j] = c * R[j, j] - s * A[j+1, j]
-        ## This is [c -s; s c] to (j+1)-th column
-        R[j:j+1, j+1] = [c -s; s c] * [R[j, j+1]; A[j+1, j+1]]
-
-        if j < n - 1
-            ## This is [c -s; s c] to (j+2)-th column, where R is still zero
-            R[j:j+1, j+2] = [-s; c] * A[j+1, j+2]
-        end
-    end
-    ## END
-    Q, R
-end
-
-A = Tridiagonal([1, 2, 3, 4], [1, 2, 3, 4, 5], [1, 2, 3, 4])
-Q, R = bandedqr(A)
-@test Q*R ≈ A
-**Problem 5(b)** Show that the determinant of a Vandermonde matrix is
-$$
-\det V = ∏_{k=1}^n ∏_{j=k+1}^n (x_j - x_j).
-$$
