+++ src/notes/III.5.OrthogonalMatrices.jmd
@@ -81,7 +81,7 @@ A 2×2 _rotation matrix_ through angle $θ$ is
 $$
 Q_θ := \begin{bmatrix} \cos θ & -\sin θ \cr \sin θ & \cos θ \end{bmatrix}
 $$
-We have $Q ∈ SO(2)$ iff $Q = Q_θ$ for some $θ ∈ ℝ$.
+We have $Q ∈ SO(2)$ if and only if $Q = Q_θ$ for some $θ ∈ ℝ$.
 
 **Proof**
 
@@ -132,7 +132,28 @@ and $\det Q = 1$.
 
 ∎
 
-
+**Example (rotating a vector)** Consider the vector
+$$
+𝐱 = \Vectt[-\sqrt{3},-1].
+$$
+We can use the proposition above to deduce the rotation matrix that rotates this
+vector to the positive real axis is:
+$$
+{1 \over \sqrt{1+3}} \begin{bmatrix} -1 & -\sqrt{3} \\ \sqrt{3} & -1 \end{bmatrix} = 
+{1 \over 2} \begin{bmatrix} -1 & -\sqrt{3} \\ \sqrt{3} & -1 \end{bmatrix}.
+$$
+Alternatively, we could determine the matrix by computing the angle of the vector via:
+$$
+θ =  {\rm atan}(-\sqrt{3}, -1) = {\rm atan}(\sqrt{3}) - π = -{2π \over 3}.
+$$
+We thus compute:
+$$
+Q_{-θ} = \begin{bmatrix}
+\cos(2π/3) & -\sin(2π/3) \\
+\sin(2π/3) & \cos(2π/3)
+\end{bmatrix} = {1 \over 2} \begin{bmatrix} -1 & -\sqrt{3} \\ \sqrt{3} & -1 \end{bmatrix}.
+$$
+∎
 
 
 ## Reflections
