+++ src/sheets/revisions.jmd
@@ -1,0 +1,4 @@
+**Problem 5(b)** Show that the determinant of a Vandermonde matrix is
+$$
+\det V = ∏_{k=1}^n ∏_{j=k+1}^n (x_j - x_j).
+$$
